Move tests to "test" sources root and add "where" function to Table
Move tests to "test" sources root and add "where" function to Table
